+++ LearnOpengl/src/Buffer/IndexBuffer.h
@@ -4,7 +4,7 @@ class IndexBuffer
 {
 public:
 
-	IndexBuffer(const void* data, size_t count);
+	IndexBuffer(const void* data, unsigned count);
 	~IndexBuffer();
 	void bind() const;
 	void unbind() const;
